Added custom exceptions

+++ Bexchange/Controllers/BookController.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
+using Bexchange.Domain.CustomExceptions;
 using Bexchange.Domain.Models;
 using Bexchange.DTOs;
 using Bexchange.Infrastructure.Repositories.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
 
 namespace Bexchange.Controllers
 {
@@ -23,77 +25,54 @@ namespace Bexchange.Controllers
         [HttpGet]
         public async Task<IActionResult> Books()
         {
-            try
-            {
-                var books = await _contentRepo.GetAllComponents();
+            var books = await _contentRepo.GetAllComponents();
 
-                return Ok(_mapper.Map<IEnumerable<BookDto>>(books));
-            }
-            catch (Exception exc)
-            {
-                return StatusCode(500, exc.Message);
-            }
+            if (books == null) 
+                throw new NotFoundException("No books here", 404);
+
+            return Ok(_mapper.Map<IEnumerable<BookDto>>(books));
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetBook(int id)
         {
-            try
-            {
-                return Ok(_mapper.Map<BookDto>(await _contentRepo.GetComponent(id)));
-            }
-            catch (Exception exc)
-            {
-                return StatusCode(500, exc.Message);
-            }
+            var book = await _contentRepo.GetComponent(id);
+
+            if (book == null) 
+                throw new NotFoundException("Book not found", (int)HttpStatusCode.NotFound);
+
+            return Ok(_mapper.Map<BookDto>(book));
         }
 
         [HttpPost]
         public async Task<IActionResult> AddBook(BookDto book)
         {
-            try
-            {
-                var newBook = _mapper.Map<Book>(book);
+            var newBook = _mapper.Map<Book>(book);
+            await _contentRepo.AddComponent(newBook);
 
-                await _contentRepo.AddComponent(newBook);
-
-                return Ok(newBook);
-                //return CreatedAtAction(nameof(_contentRepo.AddComponent), new { id = newBook.Id }, book);
-            }
-            catch (Exception exc)
-            {
-                return StatusCode(500, exc.Message);
-            }
+            return Created(Request.Path, new { newBook.Id });
         }
 
         [HttpPut]
         public async Task<IActionResult> ModifyBook(Book book)
         {
-            try
-            {
-                await _contentRepo.ModifyComponent(book);
+            if (await _contentRepo.GetComponent(book.Id) == null) 
+                throw new NotFoundException("Book not found", (int)HttpStatusCode.NotFound);
 
-                return Ok();
-            }
-            catch (Exception exc)
-            {
-                return StatusCode(500, exc.Message);
-            }
+            await _contentRepo.ModifyComponent(book);
+
+            return Ok();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteBook(int id)
         {
-            try
-            {
-                await _contentRepo.DeleteComponent(id);
+            if (await _contentRepo.GetComponent(id) == null) 
+                throw new NotFoundException("Book not found", (int)HttpStatusCode.NotFound);
 
-                return Ok($"Book with id {id} was deleted");
-            }
-            catch (Exception exc)
-            {
-                return StatusCode(500, exc.Message);
-            }
+            await _contentRepo.DeleteComponent(id);
+
+            return Ok($"Book with id {id} was deleted");
         }
     }
 }
